major refactor work. compiles but does not work.
