+++ ddtrace/example_test.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/DataDog/dd-trace-go/v2/ddtrace"
 	"github.com/DataDog/dd-trace-go/v2/ddtrace/ext"
-	"github.com/DataDog/dd-trace-go/v2/ddtrace/mocktracer"
 	"github.com/DataDog/dd-trace-go/v2/ddtrace/opentracer"
 	"github.com/DataDog/dd-trace-go/v2/ddtrace/tracer"
 
@@ -28,11 +27,11 @@ func Example_datadog() {
 	defer tracer.Stop()
 
 	// Start a root span.
-	span := tracer.StartSpan("get.data")
+	span := ddtrace.StartSpan("get.data")
 	defer span.Finish()
 
 	// Create a child of it, computing the time needed to read a file.
-	child := tracer.StartSpan("read.file", tracer.ChildOf(span.Context()))
+	child := ddtrace.StartSpan("read.file", ddtrace.ChildOf(span.Context()))
 	child.SetTag(ext.ResourceName, "test.json")
 
 	// If you are using 128 bit trace ids and want to generate the high
@@ -47,7 +46,7 @@ func Example_datadog() {
 
 	// We may finish the child span using the returned error. If it's
 	// nil, it will be disregarded.
-	child.Finish(tracer.WithError(err))
+	child.Finish(ddtrace.WithError(err))
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -66,24 +65,25 @@ func Example_opentracing() {
 	opentracing.SetGlobalTracer(t)
 }
 
-// The code below illustrates a scenario of how one could use a mock tracer in tests
-// to assert that spans are created correctly.
-func Example_mocking() {
-	// Setup the test environment: start the mock tracer.
-	mt := mocktracer.Start()
-	defer mt.Stop()
-
-	// Run test code: in this example we will simply create a span to illustrate.
-	tracer.StartSpan("test.span").Finish()
-
-	// Assert the results: query the mock tracer for finished spans.
-	spans := mt.FinishedSpans()
-	if len(spans) != 1 {
-		// fail
-		panic("expected 1 span")
-	}
-	if spans[0].OperationName() != "test.span" {
-		// fail
-		panic("unexpected operation name")
-	}
-}
+// TODO(kjn v2): Fix mocktracer
+// // The code below illustrates a scenario of how one could use a mock tracer in tests
+// // to assert that spans are created correctly.
+// func Example_mocking() {
+// 	// Setup the test environment: start the mock tracer.
+// 	mt := mocktracer.Start()
+// 	defer mt.Stop()
+//
+// 	// Run test code: in this example we will simply create a span to illustrate.
+// 	tracer.StartSpan("test.span").Finish()
+//
+// 	// Assert the results: query the mock tracer for finished spans.
+// 	spans := mt.FinishedSpans()
+// 	if len(spans) != 1 {
+// 		// fail
+// 		panic("expected 1 span")
+// 	}
+// 	if spans[0].OperationName() != "test.span" {
+// 		// fail
+// 		panic("unexpected operation name")
+// 	}
+// }
