+++ ddtrace/span_option.go
@@ -1,0 +1,112 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016 Datadog, Inc.
+
+package ddtrace
+
+import (
+	"math"
+	"time"
+
+	"github.com/DataDog/dd-trace-go/v2/ddtrace/ext"
+)
+
+// Tag sets the given key/value pair as a tag on the started Span.
+func Tag(k string, v interface{}) StartSpanOption {
+	return func(cfg *StartSpanConfig) {
+		if cfg.Tags == nil {
+			cfg.Tags = map[string]interface{}{}
+		}
+		cfg.Tags[k] = v
+	}
+}
+
+// ServiceName sets the given service name on the started span. For example "http.server".
+func ServiceName(name string) StartSpanOption {
+	return Tag(ext.ServiceName, name)
+}
+
+// ResourceName sets the given resource name on the started span. A resource could
+// be an SQL query, a URL, an RPC method or something else.
+func ResourceName(name string) StartSpanOption {
+	return Tag(ext.ResourceName, name)
+}
+
+// SpanType sets the given span type on the started span. Some examples in the case of
+// the Datadog APM product could be "web", "db" or "cache".
+func SpanType(name string) StartSpanOption {
+	return Tag(ext.SpanType, name)
+}
+
+var measuredTag = Tag(keyMeasured, 1)
+
+// Measured marks this span to be measured for metrics and stats calculations.
+func Measured() StartSpanOption {
+	// cache a global instance of this tag: saves one alloc/call
+	return measuredTag
+}
+
+// WithSpanID sets the SpanID on the started span, instead of using a random number.
+// If there is no parent Span (eg from ChildOf), then the TraceID will also be set to the
+// value given here.
+func WithSpanID(id uint64) StartSpanOption {
+	return func(cfg *StartSpanConfig) {
+		cfg.SpanID = id
+	}
+}
+
+// StartTime sets a custom time as the start time for the created span. By
+// default a span is started using the creation time.
+func StartTime(t time.Time) StartSpanOption {
+	return func(cfg *StartSpanConfig) {
+		cfg.StartTime = t
+	}
+}
+
+// AnalyticsRate sets a custom analytics rate for a span. It decides the percentage
+// of events that will be picked up by the App Analytics product. It's represents a
+// float64 between 0 and 1 where 0.5 would represent 50% of events.
+func AnalyticsRate(rate float64) StartSpanOption {
+	if math.IsNaN(rate) {
+		return func(cfg *StartSpanConfig) {}
+	}
+	return Tag(ext.EventSampleRate, rate)
+}
+
+// FinishTime sets the given time as the finishing time for the span. By default,
+// the current time is used.
+func FinishTime(t time.Time) FinishOption {
+	return func(cfg *FinishConfig) {
+		cfg.FinishTime = t
+	}
+}
+
+// WithError marks the span as having had an error. It uses the information from
+// err to set tags such as the error message, error type and stack trace. It has
+// no effect if the error is nil.
+func WithError(err error) FinishOption {
+	return func(cfg *FinishConfig) {
+		cfg.Error = err
+	}
+}
+
+// NoDebugStack prevents any error presented using the WithError finishing option
+// from generating a stack trace. This is useful in situations where errors are frequent
+// and performance is critical.
+func NoDebugStack() FinishOption {
+	return func(cfg *FinishConfig) {
+		cfg.NoDebugStack = true
+	}
+}
+
+// StackFrames limits the number of stack frames included into erroneous spans to n, starting from skip.
+func StackFrames(n, skip uint) FinishOption {
+	if n == 0 {
+		return NoDebugStack()
+	}
+	return func(cfg *FinishConfig) {
+		cfg.StackFrames = n
+		cfg.SkipStackFrames = skip
+	}
+}
