+++ ddtrace/tracer/util.go
@@ -9,54 +9,7 @@ import (
 	"fmt"
 	"strconv"
 	"strings"
-
-	"github.com/DataDog/dd-trace-go/v2/internal/samplernames"
 )
-
-// toFloat64 attempts to convert value into a float64. If the value is an integer
-// greater or equal to 2^53 or less than or equal to -2^53, it will not be converted
-// into a float64 to avoid losing precision. If it succeeds in converting, toFloat64
-// returns the value and true, otherwise 0 and false.
-func toFloat64(value interface{}) (f float64, ok bool) {
-	const max = (int64(1) << 53) - 1
-	const min = -max
-	switch i := value.(type) {
-	case byte:
-		return float64(i), true
-	case float32:
-		return float64(i), true
-	case float64:
-		return i, true
-	case int:
-		return float64(i), true
-	case int8:
-		return float64(i), true
-	case int16:
-		return float64(i), true
-	case int32:
-		return float64(i), true
-	case int64:
-		if i > max || i < min {
-			return 0, false
-		}
-		return float64(i), true
-	case uint:
-		return float64(i), true
-	case uint16:
-		return float64(i), true
-	case uint32:
-		return float64(i), true
-	case uint64:
-		if i > uint64(max) {
-			return 0, false
-		}
-		return float64(i), true
-	case samplernames.SamplerName:
-		return float64(i), true
-	default:
-		return 0, false
-	}
-}
 
 // parseUint64 parses a uint64 from either an unsigned 64 bit base-10 string
 // or a signed 64 bit base-10 string representing an unsigned integer
